Moved the header and settings buttons form the settings section into the workflow manager page. Refactored the settings page to no longer user this.parent references.
