+++ client/templates/includes/settings.pug
@@ -1,33 +1,7 @@
-div#workflow-settings.card.card-body
+div#workflow-settings
 
-  div
+  div(data-hook="timespan-settings-container")
 
-    h3.inline Settings
+  div(data-hook="param-sweep-settings-container")
 
-    button.btn.btn-outline-collapse(data-toggle="collapse" data-target="#collapse-wkfl-settings" data-hook="collapse-settings") -
-
-  div.collapse(class="show" id="collapse-wkfl-settings" data-hook="workflow-editor-container")
-
-    div(data-hook="timespan-settings-container")
-
-    div(data-hook="param-sweep-settings-container")
-
-    div(data-hook="sim-settings-container")
-
-    div.mdl-edit-btn
-
-      button.btn.btn-primary.box-shadow(data-hook="save") Save
-
-      div.mdl-edit-btn.saving-status(data-hook="saving-workflow")
-
-        div.spinner-grow
-
-        span Saving...
-
-      div.mdl-edit-btn.saved-status(data-hook="saved-workflow")
-
-        span Saved
-
-      button.btn.btn-primary.box-shadow(data-hook="start-job") Start Workflow
-      
-      button.btn.btn-primary.box-shadow(data-hook="edit-model") Edit Model
+  div(data-hook="sim-settings-container")
