+++ client/views/settings.js
@@ -16,9 +16,6 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-let $ = require('jquery');
-let path = require('path');
-let _ = require('underscore');
 //support file
 let app = require('../app');
 //views
@@ -31,96 +28,22 @@ let template = require('../templates/includes/settings.pug');
 
 module.exports = View.extend({
   template: template,
-  events: {
-    'click [data-hook=collapse-settings]' : 'changeCollapseButtonText',
-    'click [data-hook=save]' : 'clickSaveHandler',
-    'click [data-hook=start-job]'  : 'clickStartJobHandler',
-    'click [data-hook=edit-model]' : 'clickEditModelHandler'
-  },
   initialize: function (attrs, options) {
     View.prototype.initialize.apply(this, arguments);
+    this.readOnly = Boolean(attrs.readOnly) ? attrs.readOnly : false
+    this.newFormat = attrs.newFormat;
+    this.stochssModel = attrs.stochssModel;
+    this.type = attrs.type;
   },
   render: function() {
     View.prototype.render.apply(this, arguments);
-    if(this.parent.model.newFormat) {
-      $(this.queryByHook("start-job")).text("Start New Job");
+    if(this.newFormat) {
       this.renderTimespanSettingsView();
     }
-    if(this.parent.model.type === "Parameter Sweep") {
+    if(this.type === "Parameter Sweep") {
       this.renderParameterSettingsView();
-      if(this.model.parameterSweepSettings.parameters.length < 1) {
-        $(this.queryByHook("start-job")).prop("disabled", true);
-      }
-      this.model.parameterSweepSettings.parameters.on("add remove", _.bind(function (e) {
-        let numParams = this.model.parameterSweepSettings.parameters.length;
-        $(this.queryByHook("start-job")).prop("disabled", numParams < 1);
-      }, this))
     }
     this.renderSimulationSettingsView();
-  },
-  changeCollapseButtonText: function (e) {
-    app.changeCollapseButtonText(this, e);
-  },
-  clickEditModelHandler: function (e) {
-    this.parent.handleSaveWorkflow(e, _.bind(function () {
-      let queryStr = "?path=" + this.parent.model.model;
-      let endpoint = path.join(app.getBasePath(), "stochss/models/edit") + queryStr;
-      window.location.href = endpoint;
-    }, this));
-  },
-  clickSaveHandler: function (e) {
-    this.saving();
-    this.parent.handleSaveWorkflow(e, _.bind(this.saved, this));
-  },
-  clickStartJobHandler: function (e) {
-    this.saving();
-    let self = this;
-    let type = this.parent.model.type === "Ensemble Simulation" ? "gillespy" : "parameterSweep";
-    let data = {"settings": this.parent.model.settings.toJSON(),
-                "mdl_path": this.parent.model.model,
-                "type": type, "time_stamp": this.getTimeStamp()};
-    let queryStr = "?path=" + this.parent.model.directory + "&data=" + JSON.stringify(data);
-    let initEndpoint = path.join(app.getApiPath(), "workflow/init-job") + queryStr;
-    app.getXHR(initEndpoint, {
-      success: function (err, response, body) {
-        self.saved();
-        let runQuery = "?path=" + body + "&type=" + type;
-        let runEndpoint = path.join(app.getApiPath(), "workflow/run-job") + runQuery;
-        app.getXHR(runEndpoint, {
-          success: function (err, response, body) {
-            self.parent.updateWorkflow(true);
-          }
-        });
-      }
-    });
-  },
-  getTimeStamp: function () {
-    if(!this.parent.model.newFormat) {
-      return null;
-    }
-    var date = new Date();
-    var year = date.getFullYear();
-    var month = date.getMonth() + 1;
-    if(month < 10){
-      month = "0" + month
-    }
-    var day = date.getDate();
-    if(day < 10){
-      day = "0" + day
-    }
-    var hours = date.getHours();
-    if(hours < 10){
-      hours = "0" + hours
-    }
-    var minutes = date.getMinutes();
-    if(minutes < 10){
-      minutes = "0" + minutes
-    }
-    var seconds = date.getSeconds();
-    if(seconds < 10){
-      seconds = "0" + seconds
-    }
-    return "_" + month + day + year + "_" + hours + minutes + seconds;
   },
   renderParameterSettingsView: function () {
     if(this.parameterSettingsView) {
@@ -128,7 +51,8 @@ module.exports = View.extend({
     }
     this.parameterSettingsView = new ParameterSettingsView({
       model: this.model.parameterSweepSettings,
-      modelDirectory: this.parent.model.model
+      stochssModel: this.stochssModel,
+      readOnly: this.readOnly
     });
     app.registerRenderSubview(this, this.parameterSettingsView, "param-sweep-settings-container");
   },
@@ -137,7 +61,8 @@ module.exports = View.extend({
       this.simulationSettingsView.remove();
     }
     this.simulationSettingsView = new SimulationSettingsView({
-      model: this.model.simulationSettings
+      model: this.model.simulationSettings,
+      readOnly: this.readOnly
     });
     app.registerRenderSubview(this, this.simulationSettingsView, "sim-settings-container");
   },
@@ -146,16 +71,9 @@ module.exports = View.extend({
       this.timespanSettingsView.remove();
     }
     this.timespanSettingsView = new TimespanSettingsView({
-      model: this.model.timespanSettings
+      model: this.model.timespanSettings,
+      readOnly: this.readOnly
     });
     app.registerRenderSubview(this, this.timespanSettingsView, "timespan-settings-container");
-  },
-  saved: function () {
-    $(this.queryByHook('saving-workflow')).css("display", "none");
-    $(this.queryByHook('saved-workflow')).css("display", "inline-block");
-  },
-  saving: function () {
-    $(this.queryByHook('saving-workflow')).css("display", "inline-block");
-    $(this.queryByHook('saved-workflow')).css("display", "none");
   }
 });
